adjust the “country-level urban type classification” variable to [register properly to/associate with] the GEOID of the census tracks from our other datasets: Added county-level GEOID so you can join county characteristics to tracts.
change the continuous variable of “percent whiteness” to be categorical of "majority white" or "non-majority white"
(the case_when argument is not working for our data):
I don't see where you tried a case_when() function, so I don't know what went wrong, but all you need is an ifelse() if you only have two categories.

clean up the columns so that the unreferenced columns do not join with the rest of our data:
Use select() to select the columns you want to keep.